module file removed and components folder created
