+++ source/app/shared/components/footer/footer.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'app-footer',
+    templateUrl: './footer.component.html',
+    standalone: true,
+    styleUrl: './footer.component.css',
+})
+export class FooterComponent {}
