+++ source/app/shared/components/navbar/navbar.component.html
@@ -1,0 +1,43 @@
+<header
+    class="fixed top-0 w-full text-black bg-white body-font shadow-1"
+    [class.navbar-hero]="hero && !isScrolled"
+>
+    <div
+        class="container flex flex-col flex-wrap items-center p-5 mx-auto md:flex-row"
+    >
+        <a class="flex items-center mb-4 font-medium title-font md:mb-0">
+            <span class="ml-3 text-xl">Seeker</span>
+        </a>
+        <nav
+            class="flex flex-wrap items-center justify-center text-base md:mr-auto md:ml-4 md:py-1 md:pl-4 md:border-l md:border-gray-200"
+        >
+            <a
+                class="mr-5 cursor-pointer hover:text-gray-900"
+                [routerLink]="['/']"
+            >
+                Home
+            </a>
+            <a
+                class="mr-5 cursor-pointer hover:text-gray-900"
+                [routerLink]="['/about-us']"
+                >About Us</a
+            >
+        </nav>
+        <button
+            class="inline-flex items-center px-3 py-1 mt-4 text-base text-gray-600 bg-gray-100 border-0 rounded focus:outline-none hover:bg-gray-200 md:mt-0"
+        >
+            Back
+            <svg
+                fill="none"
+                stroke="currentColor"
+                stroke-linecap="round"
+                stroke-linejoin="round"
+                stroke-width="2"
+                class="w-4 h-4 ml-1"
+                viewBox="0 0 24 24"
+            >
+                <path d="M5 12h14M12 5l7 7-7 7"></path>
+            </svg>
+        </button>
+    </div>
+</header>
