--- source/app/shared/footer/footer.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'app-footer',
-    templateUrl: './footer.component.html',
-    standalone: true,
-    styleUrl: './footer.component.css',
-})
-export class FooterComponent {}
